Added component & containers folder and files added PropTypes package
