+++ index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ScoreBoard from './src/containers/ScoreBoard';
+
+ReactDOM.render(
+    <ScoreBoard />,
+    document.getElementById('root')
+  );
+
