+++ package.json
@@ -17,6 +17,7 @@
   "author": "Beau Palmquist",
   "license": "MIT",
   "dependencies": {
+    "prop-types": "^15.6.0",
     "react": "^15.3.2",
     "react-dom": "^15.3.2",
     "react-redux": "^5.0.6",
