+++ src/components/Counter.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Counter = (props) => {
+ return (
+   <div className="counter" >
+     <button className="counter-action decrement" onClick={() => props.onChange(-1)}>
+       -
+     </button>
+     <div className="counter-score"> {props.score} </div>
+     <button className="counter-action increment" onClick={() => props.onChange(1)}>
+       +
+     </button>
+   </div>
+ );
+}
+
+Counter.propTypes = {
+  onChange: PropTypes.func.isRequired,
+  score: PropTypes.number.isRequired,
+};
+
+export default Counter;
